requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
llama-parse
langchain
langchain-community
langchain-chroma
langchain-cohere
langchain-google-genai
chromadb
